Fix comments for style, destructor, curly.
