+++ examples/practice-typescript/src/components/SideBarMenu/index.scss
