+++ examples/practice-typescript/src/components/SideBarMenuItem/index.scss
@@ -1,12 +1,11 @@
 @import '../../styles/bases/variable.scss';
 
 .aside__item {
+  display: flex;
   flex-direction: column;
   justify-content: center;
   height: 60px;
   text-decoration: none;
-  display: flex;
-  font-family: Roboto;
   font-weight: 400;
   color: $color-text-secondary;
 }
@@ -18,7 +17,7 @@
 }
 
 .aside__active {
-  border-right: 3px solid $color-active-secondary;
+  border-right: 3px solid $border-right-color;
   font-weight: 700;
-  box-shadow: 3px 0px 0px 0px $color-active-secondary;
+  box-shadow: 3px 0px 0px 0px $border-right-color;
 }
