+++ examples/practice-typescript/src/styles/bases/variable.scss
@@ -15,7 +15,7 @@ $color-table: #718096;
 $color-page: rgba(203, 213, 224, 1);
 $color-active-primary: rgba(241, 110, 173, 1);
 $color-form: rgba(226, 232, 240, 1);
-$color-active-secondary: #e03386;
+$border-right-color: #e03386;
 $color-info: #4fd1c5;
 $bg-wrapper: #f5f6f7;
 $message-success: #270;
