+++ examples/practice-typescript/src/styles/modules/aside.scss
@@ -32,7 +32,7 @@
 
   &__active {
     font-weight: 700;
-    border-right: toRem(6px) solid $color-active-secondary;
+    border-right: toRem(6px) solid $border-right-color;
     margin-right: toRem(-3px);
   }
